% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{add_modules_to_network}
\alias{add_modules_to_network}
\title{Internal function for adding a set of modules to the network}
\usage{
add_modules_to_network(network, module_list)
}
\arguments{
\item{network}{The network to modify.}

\item{module_list}{A list of 'network_module' objects to add to the network.}
}
\value{
The modified network.
}
\description{
Internal function for adding a set of modules to the network
}
