% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{add_sign_to_network}
\alias{add_sign_to_network}
\title{Add random signs to network edges}
\usage{
add_sign_to_network(network)
}
\arguments{
\item{network}{The network to modify. Can be a network object or adjacency
matrix.}
}
\value{
A network object with added structure to cliques, hubs, and modules.
}
\description{
A random direction of association (positive or negative) is added to each 
connection in the cliques, hubs, and modules.
}
