% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{add_weight_to_network}
\alias{add_weight_to_network}
\title{Add random weights to network edges}
\usage{
add_weight_to_network(network, intensity = 1)
}
\arguments{
\item{network}{The network to modify.}

\item{intensity}{a positive number; this determines the scale of the weights.}
}
\value{
A network with a weighted adjacency matrix.
}
\description{
Used by the generator when generating samples; the strength of the connections
between genes varies from sample to sample. This function generates a random
strength for each edge from a truncated Normal distribution. The 
intensity parameter specifies the standard deviation of the Normal distribution.
}
