% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{all_networks_contain_same_modules}
\alias{all_networks_contain_same_modules}
\title{Internal function to check if a list of networks all contain the same modules.}
\usage{
all_networks_contain_same_modules(network_list)
}
\arguments{
\item{network_list}{A list of 'network' objects.}
}
\value{
A logical value; TRUE indicates the networks contain the same modules,
FALSE indicates otherwise. Note, this only checks that the modules contain
the same nodes - the structure of the modules are allowed to differ.
}
\description{
Internal function to check if a list of networks all contain the same modules.
}
