% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{all_networks_contain_same_nodes}
\alias{all_networks_contain_same_nodes}
\title{Internal function to check if a list of networks all contain the same nodes.}
\usage{
all_networks_contain_same_nodes(network_list)
}
\arguments{
\item{network_list}{A list of 'network' objects.}
}
\value{
A logical value; TRUE indicates the networks contain the same nodes,
FALSE indicates otherwise.
}
\description{
Internal function to check if a list of networks all contain the same nodes.
}
