% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{as.CONFUSION}
\alias{as.CONFUSION}
\title{Coerce a matrix into "CONFUSION" object}
\usage{
as.CONFUSION(mat)
}
\arguments{
\item{mat}{A 2 by 2 matrix containing integer values.}
}
\value{
An object of class "CONFUSION" containing the same counts as mat.
}
\description{
Coerces a 2 by 2 matrix containing integer counts into an object of class
"CONFUSION".
}
