% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{as.data.frame.CONFUSION}
\alias{as.data.frame.CONFUSION}
\title{Coerce "CONFUSION" object into a data.frame}
\usage{
\method{as.data.frame}{CONFUSION}(confusion, ...)
}
\arguments{
\item{confusion}{An object of class "CONFUSION".}

\item{...}{Additional arguments; these are currently ignored.}
}
\value{
Returns a data.frame object.
}
\description{
Function to coerce an object of class "CONFUSION" into a data.frame.
}
