% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{check_adjacency_cpp}
\alias{check_adjacency_cpp}
\title{C++ implementation to check if a matrix is an adjacency matrix}
\usage{
check_adjacency_cpp(m)
}
\arguments{
\item{m}{A matrix to check.}
}
\value{
Returns 0 if the matrix is an adjacency matrix. If the matrix is
not square, returns 1; if the diagonal entries are not all zero, returns 2;
if the matrix is not symmetric, returns 3; if the matrix contains values
other than 0 or 1, returns 4.
}
\description{
C++ implementation to check if a matrix is an adjacency matrix
}
