% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_adjacency_matrix}
\alias{check_adjacency_matrix}
\title{Wrapper for ArgumentCheck::addError()}
\usage{
check_adjacency_matrix(adjacency_matrix, checklist)
}
\description{
Wrapper for ArgumentCheck::addError()
}
\details{
see ?ArgumentCheck::addError() for details.
}
