% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_in_closed_interval}
\alias{check_in_closed_interval}
\title{Wrapper for ArgumentCheck::addError()}
\usage{
check_in_closed_interval(x, checklist, lower_bound, upper_bound)
}
\description{
Wrapper for ArgumentCheck::addError()
}
\details{
see ?ArgumentCheck::addError() for details.
}
