% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_positive_integer}
\alias{check_positive_integer}
\title{Wrapper for ArgumentCheck::addError()}
\usage{
check_positive_integer(p, checklist)
}
\description{
Wrapper for ArgumentCheck::addError()
}
\details{
see ?ArgumentCheck::addError() for details.
}
