% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{components_in_adjacency}
\alias{components_in_adjacency}
\title{C++ implementation to obtain connected components in a graph.}
\usage{
components_in_adjacency(adj)
}
\arguments{
\item{adj}{An adjacency matrix.}
}
\value{
Returns a matrix with 2 columns containing the indicies in the
lower-triangle of the matrix that are nonzero.
}
\description{
C++ implementation to obtain connected components in a graph.
}
