% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{create_empty_module}
\alias{create_empty_module}
\title{Create a module}
\usage{
create_empty_module(nodes)
}
\arguments{
\item{nodes}{A numeric vector indicating which nodes in the network are
contained in this module.}
}
\value{
A 'network_module' object.
}
\description{
Create a module
}
