% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{create_module_from_adjacency_matrix}
\alias{create_module_from_adjacency_matrix}
\title{Create a module from an adjacency matrix}
\usage{
create_module_from_adjacency_matrix(adjacency_matrix, nodes = NULL,
  module_name = NULL)
}
\arguments{
\item{nodes}{A numeric vector indicating which nodes in the network are
contained in this module.}

\item{module_name}{(optional) Character string specifying the name of the 
module. If NULL, the module will be unnamed.}
}
\value{
A 'network_module' object.
}
\description{
Create a module from an adjacency matrix
}
