% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R, R/networks.R
\name{create_module_from_association_matrix}
\alias{create_module_from_association_matrix}
\title{Create a module from an association matrix}
\usage{
create_module_from_association_matrix(association_matrix,
  standardize = TRUE, ...)

create_module_from_association_matrix(association_matrix,
  standardize = TRUE, ...)
}
\arguments{
\item{association_matrix}{The association matrix for the network. This is converted
to a single module structure with partial correlations specified by the
matrix.}

\item{standardize}{If True, then 'association_matrix' will be standardized
to obtain partial correlations.}

\item{...}{Additional arguments passed to 
create_module_from_association_matrix().}

\item{association}{matrix}

\item{nodes}{A numeric vector indicating which nodes in the network are
contained in this module.}

\item{module_name}{(optional) Character string specifying the name of the 
module. If NULL, the module will be unnamed.}
}
\value{
A 'network_module' object.

A network object.
}
\description{
The edge weights in the module will be set to the corresponding values
in the association matrix. Note, only the off-diagonal elements of the 
association matrix are used.
}
