% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{create_modules_for_network}
\alias{create_modules_for_network}
\title{Randomly sample subsets of genes for each module}
\usage{
create_modules_for_network(n_modules, p, avg_module_size = 20,
  sd_module_size = 10, min_module_size = 5, max_module_size = 100,
  selection_weight = 20, ...)
}
\arguments{
\item{n_modules}{The number of modules to include in the network.}

\item{p}{The number of nodes in the network.}

\item{avg_module_size}{The average number of nodes in a module.}

\item{sd_module_size}{The standard deviation of module size.}

\item{min_module_size}{The minimum number of nodes in a module.}

\item{max_module_size}{A positive value. Any generated module sizes above this 
value will be reduced to 'max_module_size'. Set to 'Inf' to avoid this 
truncation.}

\item{selection_weight}{A positive value used for sampling nodes for a new 
module.}

\item{...}{Additional arguments passed to random_module().}
}
\value{
A list containing the indicies for genes contained in each module.
}
\description{
Creates a collection of modules containing randomly samples genes.
}
