% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{create_network_from_adjacency}
\alias{create_network_from_adjacency}
\title{Create a network object from adjacency matrix}
\usage{
create_network_from_adjacency(adj_matrix, p = ncol(adj_matrix))
}
\arguments{
\item{adj_matrix}{The adjacency matrix for the network. This is converted
to a single module structure.}
}
\value{
A network object.
}
\description{
Create a network object from adjacency matrix
}
