% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{create_network_from_precision_matrix}
\alias{create_network_from_precision_matrix}
\title{Create a network object from adjacency matrix}
\usage{
create_network_from_precision_matrix(precision_matrix,
  standardize = TRUE, ...)
}
\arguments{
\item{precision_matrix}{The precision matrix for the network. This is converted
to a single module structure with partial correlations specified by the
precision matrix.}

\item{standardize}{If True, then 'precision_matrix' will be standardized
to obtain partial correlations.}

\item{...}{Additional arguments passed to 
create_module_from_precision_matrix().}
}
\value{
A network object.
}
\description{
Create a network object from adjacency matrix
}
