% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ecdf_cpp}
\alias{ecdf_cpp}
\title{C++ implementation of empirical CDF}
\usage{
ecdf_cpp(x)
}
\arguments{
\item{x}{The observation to construct the empirical CDF from.}
}
\value{
Returns the values for F(x).
}
\description{
Constructs the empirical CDF, F, for a set of observations, x, and 
returns F(x).
}
