% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edges_from_adjacency_cpp}
\alias{edges_from_adjacency_cpp}
\title{C++ implementation for obtaining an edge list from adjacency matrix}
\usage{
edges_from_adjacency_cpp(adj)
}
\arguments{
\item{adj}{An adjacency matrix.}
}
\value{
Returns a matrix with 2 columns containing the indicies in the
lower-triangle of the matrix that are nonzero.
}
\description{
C++ implementation for obtaining an edge list from adjacency matrix
}
