% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gamma_poisson.R
\name{gen_gamma_poisson}
\alias{gen_gamma_poisson}
\title{Generate sample from overdispered Poisson}
\usage{
gen_gamma_poisson(n, network, mu, overdispersion_param = 1,
  intensity_param = 1, k = 1)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{network}{The underlying network (from create_network())}

\item{intensity_param}{The standard deviation of sampled edge weights.}

\item{k}{Adjusts the tail behavior in the simulated expression distribution.}

\item{overdis_param}{The overdispersion parameters (shape, scale) used 
in the Gamma distribution.}
}
\value{
A list containing the n by p matrix of samples and the true
  association matrix.
}
\description{
Generates count data based on the overdispered Poisson distribution.
}
