% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gaussian.R
\name{gen_gaussian}
\alias{gen_gaussian}
\title{Generate sample from overdispered Poisson}
\usage{
gen_gaussian(n, network)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{network}{The underlying network (from create_network())}
}
\value{
A list containing the n by p matrix of samples, the underlying network,
and an n by p matrix containing mean expression values for each gene on each sample
after incorporating the underlying network.
}
\description{
Generates count data based on the overdispered Poisson distribution.
}
\examples{
n <- 10
network <- create_network(p = 100, modules = list(1:100)) # Create small-world network.
mu <- get_reference_count_means()
# Generate RNA-seq data using the network and reference means:
x <- gen_gaussian(n, network, mu)$x
}
