% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gaussian.R
\name{gen_gaussian_list}
\alias{gen_gaussian_list}
\title{Generate samples from a Gaussian graphical model.}
\usage{
gen_gaussian_list(n, network_list)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{network_list}{A list of networks (from [create_network()]).}
}
\value{
A list containing output from [gen_gaussian()] for each network.
}
\description{
Generates count data based on the multivariate normal distribution. Input
consists of a list of networks, for which corresponding samples are generated.
}
