% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_zinb.R
\name{gen_params_from_reference}
\alias{gen_params_from_reference}
\title{Estimate gamma-Poisson parameters from reference dataset}
\usage{
gen_params_from_reference(reference = NULL, verbose = TRUE)
}
\arguments{
\item{reference}{Either a vector or data.frame of counts from a reference
gene expression profile. If a data.frame is provided, each column should
correspond to a gene. If NULL, the kidney dataset is used.}

\item{verbose}{Boolean indicator for message output.}
}
\value{
a list containing the shape and scale parameters. If the reference
dataset is a data.frame, these parameters will be vectors containing 
estimated values for each gene.
}
\description{
The observations in the reference dataset should be as homogeneous as possible.
For example, we should not expect differential expression or differential 
connectivity of genes within the sample. If the data are heterogeneous, the
estimation of the shape and scale parameters may be unreliable.
}
