% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_adjacency_matrix.default}
\alias{get_adjacency_matrix.default}
\title{Get adjacency matrix}
\usage{
\method{get_adjacency_matrix}{default}(x, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{...}{Additional arguments.}
}
\value{
An adjacency matrix with entry ij = 1 if node i and j are 
connected, and 0 otherwise. The diagonal entries are all zero.
}
\description{
The adjacency matrix is constructed from all modules in a network.
}
