% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_adjacency_matrix.matrix}
\alias{get_adjacency_matrix.matrix}
\title{Get adjacency matrix}
\usage{
\method{get_adjacency_matrix}{matrix}(x, ...)
}
\arguments{
\item{x}{A 'matrix' object.}

\item{...}{Additional arguments.}
}
\value{
An adjacency matrix with entry ij = 1 if node i and j are 
connected, and 0 otherwise. The diagonal entries are all zero.
}
\description{
All off-diagonal entries that are near zero are set to zero. All remaining 
non-zero values are set to 1, and the diagonal is set to zero.
}
