% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_adjacency_matrix.network}
\alias{get_adjacency_matrix.network}
\title{Get adjacency matrix of a network}
\usage{
\method{get_adjacency_matrix}{network}(network, ...)
}
\arguments{
\item{network}{A 'network' object; can be either weighted or unweighted.}
}
\value{
An adjacency matrix with entry ij = 1 if node i and j are 
connected, and 0 otherwise. The diagonal entries are all zero.
}
\description{
The adjacency matrix is constructed from all modules in the network
and summarizes the connectivity among nodes.
}
