% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_adjacency_matrix.network_module}
\alias{get_adjacency_matrix.network_module}
\title{Get adjacency matrix of a module}
\usage{
\method{get_adjacency_matrix}{network_module}(x, ...)
}
\arguments{
\item{x}{The 'network_module' object to get adjacency matrix for.}

\item{...}{Additional arguments.}
}
\value{
An adjacency matrix with entry ij == 1 if node i and j are 
connected, and 0 otherwise.
}
\description{
Get adjacency matrix of a module
}
