% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{get_adjacency_matrix}
\alias{get_adjacency_matrix}
\title{Infer adjacency matrix from scores}
\usage{
get_adjacency_matrix(scores, significance = 0.1, method = 1)
}
\arguments{
\item{scores}{A symmetric p by p matrix of association scores.}

\item{significance}{False discovery rate to use.}
}
\value{
A list containing the p by p adjacency matrix and all significant 
  scores.
}
\description{
Determines which associations are significant using emperical 
Bayes for false discovery rate adjustment.
}
