% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_association_matrix.default}
\alias{get_association_matrix.default}
\title{Get association matrix}
\usage{
\method{get_association_matrix}{default}(...)
}
\arguments{
\item{...}{Either a 'network', 'network_module', or 'matrix' object.}
}
\value{
An association matrix with entry ij != 0 if node i and j are 
connected, and 0 otherwise. The diagonal entries are all zero.
}
\description{
The association matrix is constructed from all modules in a network.
}
