% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_association_matrix.matrix}
\alias{get_association_matrix.matrix}
\title{Get association matrix}
\usage{
\method{get_association_matrix}{matrix}(m, ...)
}
\arguments{
\item{m}{A 'matrix' object.}
}
\value{
Returns the matrix with diagonal elements set to zero.
}
\description{
The association matrix is constructed from all modules in a network.
}
