% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_association_matrix.network}
\alias{get_association_matrix.network}
\title{Get the association matrix of a network}
\usage{
\method{get_association_matrix}{network}(network, ...)
}
\arguments{
\item{network}{A weighted 'network' object.}
}
\value{
An association matrix with entry ij != 0 if node i and j are 
connected, and 0 otherwise. The diagonal of the matrix is set to 0.
}
\description{
The association matrix for a weighted network is recovered from its
covariance structure. The association matrix will contain ones along the
diagonal, and off-diagonal entries are interpreted as partial correlations.
}
