% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_association_matrix.network_module}
\alias{get_association_matrix.network_module}
\title{Get association matrix of a module}
\usage{
\method{get_association_matrix}{network_module}(module, ...)
}
\arguments{
\item{module}{A 'network_module' object; can be either weighted or unweighted.}
}
\value{
An association matrix with entry ij != 0 if node i and j are 
connected, and 0 otherwise. If the module is unweighted, then nonzero entries 
are set to 1. The diagonal of the association matrix is set to 1.
}
\description{
Get association matrix of a module
}
