% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{get_confusion_matrix}
\alias{get_confusion_matrix}
\title{Get performance measures for network inference}
\usage{
get_confusion_matrix(true, pred)
}
\arguments{
\item{true}{a network object or an adjacency matrix indicating the 
true underlying connections in the graph.}

\item{pred}{a matrix indicating the predicted connections
between the genes. Can be adjacency matrix or sparse association matrix 
(i.e. with non-significant associations removed); all non-zero entries are
changed to 1.}
}
\value{
The confusion matrix.
}
\description{
Computes the confusion matrix for an inferred network based on the true 
network.
}
