% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{get_degree_change}
\alias{get_degree_change}
\title{Get the change in degree for each node between two networks.}
\usage{
get_degree_change(network_1, network_2)
}
\arguments{
\item{network1}{A network object for the first network.}

\item{network2}{A network object for the second network.}
}
\value{
A vector of length p, containing the degree change for each node in 
the two networks.
}
\description{
Counts the connections to each node within each structure. Note, this
is not the same as the change in degree between the two adjacency matricies
of each network, which collapses individual structures.
}
