% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_edge_weights_from_module}
\alias{get_edge_weights_from_module}
\title{Get edge weights.}
\usage{
get_edge_weights_from_module(module)
}
\arguments{
\item{module}{The 'network_module' object to get adjacency matrix for.}
}
\value{
A vector containing the nonzero entries of module$struct.
}
\description{
Get edge weights.
}
