% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_zinb.R
\name{get_kidney_reference_data}
\alias{get_kidney_reference_data}
\title{Get reference data from kidney dataset}
\usage{
get_kidney_reference_data()
}
\value{
A data.frame containing a reference dataset with each column 
containing an observed expression profile for a gene.
}
\description{
Get reference data from kidney dataset
}
