% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{get_layout_for_modules}
\alias{get_layout_for_modules}
\title{Internal function used to create coordinates based on a set of modules}
\usage{
get_layout_for_modules(g, modules)
}
\arguments{
\item{g}{An 'igraph' object}

\item{modules}{A list containing sets of indicies indicating the nodes g that 
belong to each module}
}
\value{
A matrix of coordinates for plotting
}
\description{
Internal function used to create coordinates based on a set of modules
}
