% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{get_membership_matrix}
\alias{get_membership_matrix}
\title{Get the clustering of nodes in the network}
\usage{
get_membership_matrix(network)
}
\arguments{
\item{network}{A network object.}
}
\value{
A p by g matrix, M, where g is the number of structures in the network
and p is the number of nodes. M_ij = 1 indicates that node i is a member
of structure j. M_ij = 0 indicates non-membership.
}
\description{
Structure is synonnymous with cluster or group. The structures in the 
network may overlap, meaning a node can be a member of multiple clusters. 
This function identifies which of the nodes belong to each structure.
}
