% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggm.R
\name{get_network_arguments}
\alias{get_network_arguments}
\title{Internal function used to extract 'network' objects from argument list.}
\usage{
get_network_arguments(...)
}
\arguments{
\item{...}{The 'network' object(s) or list of networks.}
}
\value{
A list of 'network' objects.
}
\description{
Internal function used to extract 'network' objects from argument list.
}
