% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_network_characteristics}
\alias{get_network_characteristics}
\title{Characteristics of the network topology}
\usage{
get_network_characteristics(network)
}
\arguments{
\item{network}{A 'network', 'network_module', or 'matrix' object.}
}
\value{
A list containing characteristics of the network.
}
\description{
The average degree, clustering coefficient, and average path length are calculated.
}
