% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{get_network_inference_performance}
\alias{get_network_inference_performance}
\title{Get performance measures for network inference}
\usage{
get_network_inference_performance(pred, true)
}
\arguments{
\item{pred}{a matrix indicating the predicted connections
between the genes. Can be adjacency matrix or sparse association matrix 
(i.e. with non-significant associations removed); all non-zero entries are
changed to 1.}

\item{true}{a network object or an adjacency matrix indicating the 
true underlying connections in the graph.}
}
\value{
A list containing sensitivity, specificity, f1_score, tdr (true 
discovery rate), and tndr (true non-discovery rate). If the predicted network
is empty, then tdr is set to NA. Similarly, if the predicted network is completely
connected, tndr is set to NA.
}
\description{
Computes the sensitivity, specificity, f1score, tdr, and tndr
  to measure the accuracy of an inferred network on the true network.
}
