% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_network_modules}
\alias{get_network_modules}
\title{Get a list of modules from the network}
\usage{
get_network_modules(network)
}
\arguments{
\item{network}{A 'network' object.}
}
\value{
A list whose length is the number of modules in the network; 
each element is a vector containing the indicies of the nodes
that belong to that module.
}
\description{
Get a list of modules from the network
}
