% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_node_names.matrix}
\alias{get_node_names.matrix}
\title{Get node names}
\usage{
\method{get_node_names}{matrix}(m, ...)
}
\arguments{
\item{m}{A 'matrix' object.}
}
\value{
A vector containing the node names or node indicies.
}
\description{
Get node names
}
