% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_node_names.network_module}
\alias{get_node_names.network_module}
\title{Get node names of a module}
\usage{
\method{get_node_names}{network_module}(x, ...)
}
\arguments{
\item{x}{The 'network_module' object to get node names from.}

\item{...}{Additional arguments.}
}
\value{
A vector containing the node indicies in the module
}
\description{
Modules do not retain the names of each node, so the node indicies returned
used instead. If the network node names are known, then the vector returned
from this function can be used to index those node names.
}
