% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getter.R
\name{get_pathway_list}
\alias{get_pathway_list}
\title{Get a list of pathways in the network}
\usage{
get_pathway_list(network)
}
\arguments{
\item{network}{A network object.}
}
\value{
A list of g vectors of integers, where g is the number of pathways 
in the network and each vector contains the indicies of the nodes contained
in the pathway.
}
\description{
Pathways refer to hubs, modules, and cliques. The pathways in the 
network may overlap, meaning a node can be a member of multiple pathways. 
This function identifies which of the nodes belong to each pathway.
}
