% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gamma_poisson.R
\name{get_reference_count_means}
\alias{get_reference_count_means}
\title{Obtain empirical distribution of mean RNA-seq counts}
\usage{
get_reference_count_means(reference_dataset = NULL, min_mean_count = 30,
  max_mean_count = NULL)
}
\arguments{
\item{reference_dataset}{A reference dataset of RNA-seq expression counts. 
Can be a matrix or data.frame object.}

\item{min_mean_count}{Average counts below this value will be excluded.}

\item{max_mean_count}{Average counts above this value will be excluded.}
}
\value{
A vector containing the average count for each gene.
}
\description{
The average expression level of each gene in a reference dataset are 
used as an empirical distribution. If no reference dataset is provided,
the kidney data from the SimSeq package are used.
}
