% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_sigma.matrix}
\alias{get_sigma.matrix}
\title{Get the covariance matrix}
\usage{
\method{get_sigma}{matrix}(m, ...)
}
\arguments{
\item{m}{A 'matrix' object.}
}
\value{
A covariance matrix.
}
\description{
The associations in each module are taken as partial correlations, and
the covariance matrix is calculated from these assuming that expression
for gene i is the weighted average over each module using 1/sqrt(m_i) 
as the weight, where m_i is the number of modules containing gene i.
}
\note{
The matrix is assumed to be a partial correlation matrix.
A warning is given in this case. To avoid the warning message, convert the
matrix into a network object using 'create_network_from_association_matrix()'.
}
