% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_sigma.network}
\alias{get_sigma.network}
\title{Get the covariance matrix of a network}
\usage{
\method{get_sigma}{network}(x, ...)
}
\arguments{
\item{...}{Additional arguments.}

\item{network}{A weighted 'network' object.}
}
\value{
A covariance matrix for the network.
}
\description{
The associations in each module are taken as partial correlations, and
the covariance matrix is calculated from these assuming that expression
for gene i is the weighted average over each module using 1/sqrt(m_i) 
as the weight, where m_i is the number of modules containing gene i.
}
