% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_sigma.network_module}
\alias{get_sigma.network_module}
\title{Get the covariance matrix of a module}
\usage{
\method{get_sigma}{network_module}(module, ...)
}
\arguments{
\item{module}{A weighted 'network_module' object.}
}
\value{
A covariance matrix for the module
}
\description{
The associations in the module are taken as partial correlations; a 
precision matrix is obtained using the negative partial correlations, and
the covariance matrix is computed by inverting the precision matrix.
}
