% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gaussian.R
\name{get_sigma_from_precision}
\alias{get_sigma_from_precision}
\title{Compute the covariance matrix from a precision matrix}
\usage{
get_sigma_from_precision(precision, k = 1)
}
\arguments{
\item{precision}{The precision matrix.}

\item{k}{An integer that ensures the matrix inverse is numerically stable. 
k = 1 is default; higher values will give less stable results.}
}
\value{
A p by p covariance matrix.
}
\description{
If the precision matrix is singular, it is made invertible by adding a 
diagonal matrix.
}
