% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gaussian.R
\name{get_sigma_from_precision_list}
\alias{get_sigma_from_precision_list}
\title{Compute covariance matrices from a list of precision matrices}
\usage{
get_sigma_from_precision_list(precision_list)
}
\arguments{
\item{precision_list}{A list of precision matricies.}
}
\value{
A list of covariance matrices.
}
\description{
Compute covariance matrices from a list of precision matrices
}
