% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{get_summary_for_node}
\alias{get_summary_for_node}
\title{Get summary for a node in the network.}
\usage{
get_summary_for_node(network, node)
}
\arguments{
\item{network}{A network object.}

\item{node}{The node to summarize. Can be a character string if the nodes
are labeled, or a integer value from 1 to p.}
}
\value{
A list containing summary, a list of strings detailing each 
structure the node belongs to; struct_count, the number of structures
the node belongs to; and degree, the total number of connections to the 
node (summed over all structures).
}
\description{
Get summary for a node in the network.
}
