% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_adjacency_cpp}
\alias{is_adjacency_cpp}
\title{C++ implementation to check if a matrix is an adjacency matrix}
\usage{
is_adjacency_cpp(m)
}
\arguments{
\item{m}{A matrix to check.}
}
\value{
Returns TRUE if the matrix is an adjacency matrix and FALSE 
otherwise.
}
\description{
C++ implementation to check if a matrix is an adjacency matrix
}
