% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{is_PD}
\alias{is_PD}
\title{Check if a matrix is positive definite}
\usage{
is_PD(x)
}
\arguments{
\item{x}{A matrix to check.}
}
\value{
Returns TRUE if the matrix is positive definite and FALSE otherwise.
}
\description{
Check if a matrix is positive definite
}
