% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_symmetric_cpp}
\alias{is_symmetric_cpp}
\title{C++ implementation to check if a matrix is symmetric}
\usage{
is_symmetric_cpp(m, tol = 1e-12)
}
\arguments{
\item{m}{A matrix to check.}

\item{tol}{A Numeric scalar >= 0. Differences smaller than tol are ignored.}
}
\value{
Returns TRUE if the matrix is symmetric and FALSE otherwise.
}
\description{
C++ implementation to check if a matrix is symmetric
}
