% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is_weighted.matrix}
\alias{is_weighted.matrix}
\title{Check if an object is weighted}
\usage{
\method{is_weighted}{matrix}(x, ...)
}
\arguments{
\item{x}{A 'matrix' object.}

\item{...}{Additional arguments.}
}
\value{
Returns FALSE if all of entries in the matrix are either zero or one, 
and returns TRUE otherwise. However, if there are no connections, i.e. all
off-diagonal entries are zero, then this function returns TRUE.
}
\description{
Check if an object is weighted
}
