% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{is_weighted.network_module}
\alias{is_weighted.network_module}
\title{Check if a module is weighted}
\usage{
\method{is_weighted}{network_module}(module, ...)
}
\arguments{
\item{module}{The 'network_module' object to check.}
}
\value{
A boolean value that is FALSE if all of the connections in the 
module are weighted by 0s and 1s, and it returns TRUE otherwise. If there
are no connections in the module, then this function returns TRUE.
}
\description{
Check if a module is weighted
}
