% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is_weighted}
\alias{is_weighted}
\title{Check if an object is weighted}
\usage{
is_weighted(x, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{...}{Additional arguments.
object are weighted by 0s and 1s, and returns TRUE otherwise. If there
are no connections in the module, then this function returns TRUE.}
}
\description{
Check if an object is weighted
}
