% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{isSymmetric.shrinkage}
\alias{isSymmetric.shrinkage}
\title{Test if a shrinkage object is symmetric.}
\usage{
\method{isSymmetric}{shrinkage}(x, ...)
}
\arguments{
\item{x}{An object of class 'shrinkage'.}

\item{...}{Further argumnts passed to [isSymmetric()].}
}
\value{
Logical indicated if x is symmetric or not.
}
\description{
Test if a shrinkage object is symmetric.
}
