% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reference.R
\name{load_reference_data}
\alias{load_reference_data}
\title{Load the RNA-seq reference data}
\usage{
load_reference_data()
}
\value{
A list containing the RNA-seq reference data and the corresponding 
ZINB parameters.
}
\description{
Load the RNA-seq reference data
}
