% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{new_checklist}
\alias{new_checklist}
\title{Wrapper for ArgumentCheck::newArgCheck()}
\usage{
new_checklist()
}
\description{
Wrapper for ArgumentCheck::newArgCheck()
}
\details{
see ?ArgumentCheck::newArgCheck() for details.
}
