% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpls.R
\name{normalize_cpls_scores}
\alias{normalize_cpls_scores}
\title{Normalize the cPLS scores}
\usage{
normalize_cpls_scores(scores)
}
\arguments{
\item{scores}{The association scores from the cPLS algorithm.}
}
\description{
Performs the transformation: sign(x) log(|x| + 1). This normalizes the
cPLS scores; should be performed prior to using `fdr()`.
}
