% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overwrite_as_adjacency_cpp}
\alias{overwrite_as_adjacency_cpp}
\title{C++ implementation to convert a matrix to an adjacency matrix}
\usage{
overwrite_as_adjacency_cpp(m, tol = 10^-13)
}
\arguments{
\item{m}{A matrix to convert.}

\item{tol}{A numeric value. Entires within tol from zero are set to zero.}
}
\value{
Returns the adjacency matrix.
}
\description{
Sets all nonzero values to 1.
}
