% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overwrite_with_qzinb_cpp}
\alias{overwrite_with_qzinb_cpp}
\title{C++ implementation of qzinb}
\usage{
overwrite_with_qzinb_cpp(p, size, mu, rho, lower = TRUE, log = FALSE)
}
\arguments{
\item{p}{A vector of probabilities}

\item{size}{The distribution mean.}

\item{mu}{The dispersion paramater used in dnbinom.}

\item{rho}{The zero-inflation parameter.}

\item{lower}{Logical; if TRUE, then probabilities are P(X <= x). 
Otherwise, P(X > x).}

\item{log}{Logical; if TRUE, then exp(p) is used.}
}
\description{
C++ implementation of qzinb
}
