% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{perturb_network}
\alias{perturb_network}
\title{Perturbs the connections in a network}
\usage{
perturb_network(network, rewire_hub_prob = 0.5,
  rewire_other_prob = 0.1)
}
\arguments{
\item{network}{The network to modify.}

\item{rewire_hub_prob}{The probability that any given hub node will have its
connections removed.}

\item{rewire_other_prob}{The probability that any given non-hub node will be
perturbed by rewiring. If a node is rewired, it will have, on average, the
same number of connections as its original state.}
}
\value{
The modified network.
}
\description{
By default, hub nodes have a 50% chance of being turned "off", and all other
nodes have a 10% chance of being perturbed by rewiring. No new hub nodes are
be created.
}
