% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{plot.CONFUSION}
\alias{plot.CONFUSION}
\title{Plots confusion matrix}
\usage{
\method{plot}{CONFUSION}(confusion, main = NULL, ...)
}
\arguments{
\item{confusion}{An object of class "CONFUSION".}

\item{main}{Title for the plot.}

\item{...}{Other graphics parameters; these are currently ignored.}
}
\value{
The generated ggplot object is returned. This can be used for further 
editing.
}
\description{
Plot method for objects of class "CONFUSION".
}
\note{
The gradient used to color each cell is proportional to the sum of
adjacent cells with the current cell; for example, the gradient to color 
the TP cell is TP / (TP + FN + FP). This re-weighting helps reveal good 
performance in the case of imbalanced class labels.
}
