% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.network_plot}
\alias{plot.network_plot}
\title{Plot function for 'network_plot' class}
\usage{
\method{plot}{network_plot}(x, ...)
}
\arguments{
\item{x}{A 'network_plot' object obtained from plot.network() or
plot_network().}

\item{...}{Additional arguments passed to plot.igraph().}
}
\description{
Plot function for 'network_plot' class
}
