% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_network_matrix}
\alias{plot_network_matrix}
\title{Plot matrix representation of network}
\usage{
plot_network_matrix(network, main = "Untitled",
  col = colorRampPalette(RColorBrewer::brewer.pal(8, "Greys"))(50))
}
\arguments{
\item{network}{Either a network object or association matrix of the network.}

\item{main}{A string containing the title for the graph.}

\item{col}{Color palatte used for heatmap. See ?heatmap for details.}
}
\value{
The matrix used to create the heatmap
}
\description{
This function plots the given network as a heatmap to visualize the 
corresponding association matrix. A weighted association matrix can be 
provided to show relative strength of associations.
}
