% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{populate_modules_for_network}
\alias{populate_modules_for_network}
\title{Randomly sample subsets of genes for each module}
\usage{
populate_modules_for_network(n_modules, p, min_module_size = 10,
  max_module_size = Inf, avg_module_size = 25, sd_module_size = 10,
  repeated_module_factor = 0.2, ...)
}
\arguments{
\item{n_modules}{The number of modules to include in the network.}

\item{p}{The number of nodes in the network.}

\item{min_module_size}{The minimum number of nodes in a module.}

\item{max_module_size}{(optional) If specified, any generated module sizes 
above this value will be reduced to 'max_module_size'.}

\item{avg_module_size}{The average number of nodes in a module.}

\item{sd_module_size}{The standard deviation of module size.}

\item{repeated_module_factor}{A value between 0 and 1. A gene selected for a 
module has its probability of being selected for another module multiplied by 
this factor.}
}
\value{
A list containing the indicies for genes contained in each module.
}
\description{
Creates a collection of modules containing randomly samples genes.
}
