% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{print.CONFUSION}
\alias{print.CONFUSION}
\title{Print function for objects of class "CONFUSION"}
\usage{
\method{print}{CONFUSION}(confusion, ...)
}
\arguments{
\item{confusion}{An object of class "CONFUSION".}

\item{...}{Additional arguments to be passed into the print function.}
}
\description{
Print the confusion matrix.
}
