% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{print.network}
\alias{print.network}
\title{Print function for 'network' object.}
\usage{
\method{print}{network}(x, ...)
}
\arguments{
\item{x}{A 'network' object.}

\item{...}{Additional arguments are ignored.}
}
\value{
Prints a summary of the module.
}
\description{
Print function for 'network' object.
}
