% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{print.network_module}
\alias{print.network_module}
\title{Print function for 'network_module' object.}
\usage{
\method{print}{network_module}(x, ...)
}
\arguments{
\item{x}{A 'network_module' object.}

\item{...}{Additional arguments are ignored.}
}
\value{
Prints a summary of the module.
}
\description{
Print function for 'network_module' object.
}
\examples{
module <- random_module(1:10)
module
print(module)
}
