% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.network_plot}
\alias{print.network_plot}
\title{Print function for 'network_plot' class}
\usage{
\method{print}{network_plot}(x, ...)
}
\arguments{
\item{x}{A 'network_plot' object obtained from \code{\link{plot.network}} or
\code{\link{plot_network}}.}

\item{...}{Additional arguments passed to \code{\link[graphics:plot.default]{plot}}.}
}
\value{
Creates a plot of the network and returns a graph object.
See \code{\link{plot_network}} for details.
}
\description{
Displays the network plot.
}
\examples{
nw <- random_network(10)
g <- plot(nw, display_plot = FALSE) # Doesn't display the plot.
g # Displays the plot.
}
