% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{random_module}
\alias{random_module}
\title{Create a random module}
\usage{
random_module(nodes, module_name = NULL, ...)
}
\arguments{
\item{nodes}{A numeric vector indicating which nodes in the network 
are contained in this module.}

\item{module_name}{(optional) Character string specifying the name of the 
module. If NULL, the module will be unnamed.}

\item{...}{Additional arguments passed to 'random_module_structure()'.}
}
\value{
A 'network_module' object.
}
\description{
Create a random module
}
\examples{
module <- random_module(1:10)
}
