% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gaussian.R
\name{random_precision_from_network}
\alias{random_precision_from_network}
\title{Generate a precision matrix from a network}
\usage{
random_precision_from_network(network)
}
\arguments{
\item{network}{The underlying network (from create_network()).}
}
\value{
A p by p precision matrix.
}
\description{
A random precision matrix (i.e. an inverse covariance matrix) is generated
based on the structure from a given network.
}
