% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gaussian.R
\name{random_precision_from_network_list}
\alias{random_precision_from_network_list}
\title{Generate partial correlations for a list of networks.}
\usage{
random_precision_from_network_list(network_list, k = 1, limits = c(0.5, 1))
}
\arguments{
\item{network_list}{A list of network objects (from [create_network()]).}

\item{k}{An integer that ensures the matrix inverse is numerically stable. 
k = 1 is default; higher values will give less stable results.}

\item{limits}{A vector of length 2 containing the lower and upper limits
for partial correlations (generated from a uniform distribution).}
}
\value{
A list of matrices containing partial correlations.
}
\description{
Matrices containing partial correlations are generated based on the structure 
of each network. Edges that are common across networks are given the same 
partial correlation value.
}
