% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference}
\alias{reference}
\title{RNA-seq reference dataset}
\format{
A list containing two matrices:
\describe{
\item{$rnaseq}{A 200 by 2000 matrix containing the raw RNA-seq
expression counts}
\item{$params}{A 3 by 2000 matrix containing the estimated
ZINB parameters for each expression profile}
}
}
\source{
\url{http://www.linkedomics.org/data_download/TCGA-BRCA/}
}
\usage{
reference
}
\description{
The reference is a breast invasive carcinoma dataset containing gene
expression profiles generated by The Cancer Genome Atlas (TCGA) and
downloaded using the LinkedOmics portal. The dataset contains 200 samples
and 2000 genes, randomly subset from the full TCGA-BRCA cohort to keep
the package size small. The reference is a list containing a matrix of
the expression data and a matrix of estimated ZINB parameters for each
expression profile.
}
\keyword{datasets}
