% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{remove_connections.default}
\alias{remove_connections.default}
\title{Remove connections in a network.}
\usage{
\method{remove_connections}{default}(...)
}
\arguments{
\item{...}{A 'matrix' object to modify.}
}
\value{
The modified object.
}
\description{
Remove connections in a network.
}
