% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{remove_connections_to_node.matrix}
\alias{remove_connections_to_node.matrix}
\title{Remove connections to a node.}
\usage{
\method{remove_connections_to_node}{matrix}(x, node, prob_remove,
  weights = NULL, ...)
}
\arguments{
\item{x}{An adjacency matrix to modify.}

\item{node}{The node to unwire.}

\item{prob_remove}{A value between 0 and 1. Each connection to 'node_index' 
will be removed with probability equal to 'prob_remove'.}

\item{weights}{(Optional) A vector of weights for each node. These are used
in addition to the degree of each node when sampling neighbors to unwire from.}

\item{...}{Additional arguments.}
}
\value{
The modified adjacency matrix.
}
\description{
Remove connections to a node.
}
