% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{remove_connections_to_node}
\alias{remove_connections_to_node}
\title{Remove node from the network}
\usage{
remove_connections_to_node(network, node)
}
\arguments{
\item{network}{The network to modify.}

\item{node}{The node to remove. Can be a character string if the nodes
are labeled, or a integer value from 1 to p.}
}
\value{
The modified network.
}
\description{
Remove node from the network
}
