% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{remove_node_from_clique}
\alias{remove_node_from_clique}
\title{Remove node from a clique in the network}
\usage{
remove_node_from_clique(network, clique_index, node_index)
}
\arguments{
\item{network}{The network to modify.}

\item{clique_index}{The index of the clique in the network to modify.}

\item{node_index}{The index of the node in the hub to remove.}
}
\value{
The modified network.
}
\description{
Remove node from a clique in the network
}
