% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{remove_node_from_hub}
\alias{remove_node_from_hub}
\title{Remove node from a hub in the network}
\usage{
remove_node_from_hub(network, hub_index, node_index)
}
\arguments{
\item{network}{The network to modify.}

\item{hub_index}{The index of the hub in the network to modify.}

\item{node_index}{The index of the node in the hub to remove. If the index
is 1, which is the hub node, the entire hub is removed from the network.}
}
\value{
The modified network.
}
\description{
Remove node from a hub in the network
}
