% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{remove_node_from_module}
\alias{remove_node_from_module}
\title{Remove node from a module in the network}
\usage{
remove_node_from_module(network, module_index, node_index)
}
\arguments{
\item{network}{The network to modify.}

\item{module_index}{The index of the module in the network to modify.}

\item{node_index}{The index of the node in the hub to remove.}
}
\value{
The modified network.
}
\description{
Remove node from a module in the network
}
