% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{remove_small_components_from_module}
\alias{remove_small_components_from_module}
\title{Removes small components within a network module.}
\usage{
remove_small_components_from_module(network, module)
}
\arguments{
\item{network}{The network to modify.}

\item{module}{The module to make connected.}
}
\value{
The modified network.
}
\description{
Connectivity of the network module is enforced by removing all but the 
largest connected component in the module.
}
