% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{remove_weights.network}
\alias{remove_weights.network}
\title{Removes the weights of all connections in the network}
\usage{
\method{remove_weights}{network}(network)
}
\arguments{
\item{network}{The 'network' object to modify.}
}
\value{
The modified 'network' object.
}
\description{
Removes the weights of all connections in the network
}
