% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{remove_weights.network_module}
\alias{remove_weights.network_module}
\title{Removes the connection weights for a module}
\usage{
\method{remove_weights}{network_module}(x, ...)
}
\arguments{
\item{x}{The 'network_module' object to modify.}

\item{...}{Additional arguments.}
}
\value{
The modified 'network_module' object.
}
\description{
Removes the connection weights for a module
}
