% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{replace_module_in_network}
\alias{replace_module_in_network}
\title{Replaces a module in the network}
\usage{
replace_module_in_network(module_index, module, network)
}
\arguments{
\item{module_index}{The index of the module to replace.}

\item{module}{The new module to replace with.}

\item{network}{The network to modify.}
}
\value{
The modified network.
}
\description{
Replaces a module in the network
}
