% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{report_checks}
\alias{report_checks}
\title{Wrapper for ArgumentCheck::finishArgCheck}
\usage{
report_checks(check)
}
\arguments{
\item{check}{An object of class ArgCheck.}
}
\description{
Wrapper for ArgumentCheck::finishArgCheck
}
\details{
see ?ArgumentCheck::finishArgCheck for details.
}
