% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rewire_connections.matrix}
\alias{rewire_connections.matrix}
\title{Rewire connections.}
\usage{
\method{rewire_connections}{matrix}(x, prob_rewire, weights = NULL,
  alpha = 100, beta = 1, epsilon = 10^-5, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{prob_rewire}{A value between 0 and 1. The connections to each node 
will be rewired with probability equal to 'prob_rewire'.}

\item{weights}{(Optional) A vector of weights for each node. These are used
in addition to the degree of each node when sampling a node to rewire to.}

\item{alpha}{A positive value used to parameterize the Beta distribution.}

\item{beta}{A positive value used to parameterize the Beta distribution.}

\item{epsilon}{A small constant added to the sampling probability of each node.}

\item{...}{Additional arguments.}
}
\value{
The modified object x.
}
\description{
Rewire connections.
}
