% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rewire_connections}
\alias{rewire_connections}
\title{Rewire connections.}
\usage{
rewire_connections(x, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{...}{Additional arguments.}
}
\value{
The modified object.
}
\description{
Rewire connections.
}
