% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{rewire_connections_to_node.network}
\alias{rewire_connections_to_node.network}
\title{Rewire connections to a node.}
\usage{
\method{rewire_connections_to_node}{network}(x, node, prob_rewire,
  weights = NULL, alpha = 100, beta = 1, epsilon = 10^-5, ...)
}
\arguments{
\item{x}{A 'network' object to modify.}

\item{node}{The node to rewire.}

\item{prob_rewire}{A value between 0 and 1. Each connection to 'node' 
will be rewired with probability equal to 'prob_rewire'. Note, the degree of 
'node' is unchanged after this operation.}

\item{weights}{(Optional) A vector of weights for each node. These are used
in addition to the degree of each node when sampling nodes to rewire.}

\item{alpha}{A positive value used to parameterize the Beta distribution.}

\item{beta}{A positive value used to parameterize the Beta distribution.}

\item{epsilon}{A small constant added to the sampling probability of each node.}

\item{...}{Additional arguments.}
}
\value{
The modified network.
}
\description{
Rewire connections to a node.
}
