% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{rewire_connections_to_node.network_module}
\alias{rewire_connections_to_node.network_module}
\title{Rewire connections to a node.}
\usage{
\method{rewire_connections_to_node}{network_module}(module, node,
  rewire_prob, weights = NULL, exponent = 0)
}
\arguments{
\item{module}{A 'network_module' object to modify.}

\item{node}{The node to rewire.}

\item{rewire_prob}{A value between 0 and 1. Each connection to 'node' 
will be rewired with probability equal to 'rewire_prob'. Note, the degree of 
'node' is unchanged after this operation.}

\item{weights}{(Optional) A vector of weights for each node. These are used
in addition to the degree of each node when sampling nodes to rewire.}

\item{exponent}{The exponent used for weighted sampling. When exponent = 0,
nodes are sampled uniformly. When exponent > 0, the sampling probability
is based on node weights.}
}
\value{
The modified module.
}
\description{
Rewire connections to a node.
}
