% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{rewire_connections_to_node}
\alias{rewire_connections_to_node}
\title{Rewire node in the network.}
\usage{
rewire_connections_to_node(node, network, rewire_prob = 0.2)
}
\arguments{
\item{node}{The node to rewire. Can be a character string 
corresponding to a name of a node in the network, or an integer value from 
1 to p corresponding to the index of a node.}

\item{network}{The network to modify.}

\item{rewire_prob}{The node at 'node_index' has probaility equal to 
'rewire_prob' to be connected with any other node in the module.}
}
\value{
The modified network.
}
\description{
Within each module the node is a member of, all connections are reset. 
New connections are made with a specified probablility.
}
