% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{rewire_connections_to_node_in_module}
\alias{rewire_connections_to_node_in_module}
\title{Rewire connections to a node from a module}
\usage{
rewire_connections_to_node_in_module(node, module, rewire_prob = 0.2)
}
\arguments{
\item{module}{The 'network_module' object to modify.}

\item{rewire_prob}{The node at 'node_index' has probaility equal to 
'rewire_prob' to be connected with any other node in the module.}

\item{node_index}{The index of the node in the module to remove.}
}
\value{
The modified 'network_module' object.
}
\description{
Rewire connections to a node from a module
}
