% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ring_lattice_cpp}
\alias{ring_lattice_cpp}
\title{C++ implementation for creating a ring lattice}
\usage{
ring_lattice_cpp(p, neig_size)
}
\arguments{
\item{p}{The number of nodes in the lattice.}

\item{neig_size}{The neighborhood side within which nodes are connected.}
}
\value{
Returns the adjacency matrix for the ring lattice.
}
\description{
C++ implementation for creating a ring lattice
}
