% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpls.R
\name{run_cpls_parallel}
\alias{run_cpls_parallel}
\title{Count data partial least squares}
\usage{
run_cpls_parallel(x, v = 3, threshold = NULL)
}
\arguments{
\item{x}{the n by p matrix of raw count data. Rows should correspond to
samples, and columns should correspond to genes.}

\item{v}{the number of PLS components to compute.}
}
\value{
a matrix of association scores.
}
\description{
Parallel version of run_cpls(). Reconstructs a full genomic 
  association network from a set of next-generation sequencing count data.
}
