% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso.R
\name{run_glasso}
\alias{run_glasso}
\title{Wrapper for glasso method}
\usage{
run_glasso(x, method = c("mb", "glasso", "ct"), criterion = c("ric",
  "stars"), verbose = FALSE, ...)
}
\arguments{
\item{x}{The n by p matrix of counts.}

\item{method}{The method used to compute correlations. Should be either "pearson"
or "spearman". The default is Spearman, which provides 
the more conservative estimation of associations.}

\item{threshold}{Cutoff for significant associations. If NULL, all scores
are returned. Otherwise, scores at or below this threshold are set to zero.}
}
\value{
A list containing the p by p matrix of association scores, and the 
threshold used to determine significant associations.
}
\description{
Conducts co-expression analysis using glasso method. P
}
