% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icorr.R
\name{run_icorr}
\alias{run_icorr}
\title{Wrapper for inverse covariance matrix}
\usage{
run_icorr(x, threshold = NULL, k = 1)
}
\arguments{
\item{x}{The n by p matrix of observations.}

\item{threshold}{Cutoff for significant associations. If NULL, all correlations
are returned. Otherwise, correlations of magnitude at or below this threshold are 
set to zero.}

\item{k}{An integer that ensures the matrix inverse is numerically stable. 
k = 1 is default; higher values will give less stable results.}
}
\value{
A list containing the p by p matrix of partial correlations and
the threshold used to determine significant associations.
}
\description{
Conducts co-expression analysis using full partial correlations; these are
computed by inverting the sample correlation matrix of the observations.
}
