% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcor.R
\name{run_pcor}
\alias{run_pcor}
\title{Wrapper for partial correlations from corpcor}
\usage{
run_pcor(x, threshold = NULL, verbose = FALSE)
}
\arguments{
\item{x}{The n by p matrix of counts.}

\item{threshold}{Cutoff for significant associations. If NULL, all correlations
are returned. Otherwise, correlations of magnitude at or below this threshold are 
set to zero.}

\item{verbose}{If true, progress messages will be printed from pcor.shrink().}
}
\value{
A list containing the p by p matrix of partial correlations and
the threshold used to determine significant associations.
}
\description{
Conducts co-expression analysis using full partial correlations; these are
computed using the shrinkage approach for covariance estimation from the 
corpcor package.
}
