% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{run_wgcna}
\alias{run_wgcna}
\title{Wrapper for WGCNA method}
\usage{
run_wgcna(x, threshold = 0, beta = 6)
}
\arguments{
\item{x}{The n by p matrix of counts}

\item{threshold}{Cutoff for significant associations}

\item{beta}{Tuning parameter for WGCNA}
}
\value{
A list containing `scores`, a p by p matrix of association scores, and 
`adj_matrix`, a p by p adjacency matrix.
}
\description{
Conducts co-expression analysis using WGCNA method.
}
