% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{sample_link_nodes}
\alias{sample_link_nodes}
\title{Sample link nodes for new module}
\usage{
sample_link_nodes(nodes, degree)
}
\arguments{
\item{nodes}{The nodes to sample from.}

\item{degree}{The degree of each node.}
}
\value{
A vector of selected nodes (possibly of length 1).
}
\description{
Sample link nodes for new module
}
