% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{sample_module_nodes}
\alias{sample_module_nodes}
\title{Sample nodes for new module}
\usage{
sample_module_nodes(nodes, m, degree, index_in_modules, index_link_nodes)
}
\arguments{
\item{nodes}{The nodes available to sample from.}

\item{m}{The number of nodes to sample.}

\item{degree}{The degree of each node.}

\item{index_in_modules}{The indicies of nodes that already belong to other 
modules. This is used to index into 'nodes' and 'degree'.}

\item{index_link_nodes}{The indicies of link nodes already chosen to
be in this module. This is used to index into 'nodes' and 'degree'.}
}
\value{
A vector of selected nodes of length m.
}
\description{
Sample nodes for new module
}
