% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{scores_to_adjacency_matrix}
\alias{scores_to_adjacency_matrix}
\title{Infer adjacency matrix from scores}
\usage{
scores_to_adjacency_matrix(scores, significance = 0.05, method = "fdr")
}
\arguments{
\item{scores}{A symmetric p by p matrix of association scores.}

\item{significance}{Must be specified if method = "fdr". 
Specifies the false discovery rate threshold to use.}

\item{method}{Character string specifing which method to use. Currently only
"fdr" is implemented.}
}
\value{
The p by p adjacency matrix containing only 1's and 0's.
}
\description{
Determines which associations are significant based on an emperical 
Bayes false discovery rate.
}
