% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_low_rank_scores_to_zero.R
\name{set_low_rank_scores_to_zero}
\alias{set_low_rank_scores_to_zero}
\title{Set low ranking scores to zero.}
\usage{
set_low_rank_scores_to_zero(scores, rank = 100, verbose = FALSE)
}
\arguments{
\item{scores}{The matrix of association scores.}

\item{rank}{Scores below this rank are set to zero.}

\item{verbose}{If true, the number of associations that were set to zero 
are printed.}
}
\value{
The matrix of association scores with low ranking scores set to zero. 
Note, in the case of ties (many edges having the same association), the total
number of edges may exceed rank.
}
\description{
The association scores are ranked (in magnitude) from strongest to weakest, 
and the bottom ranking scores are set to zero. This is useful when plotting,
if it is desired to show only the edges with the strongest associations.
}
