% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{set_module_name}
\alias{set_module_name}
\title{Set the name for a module}
\usage{
set_module_name(module, module_name)
}
\arguments{
\item{module}{The 'network_module' object to modify.}

\item{module_name}{A character string.}
}
\value{
The modified 'network_module' object.
}
\description{
Set the name for a module
}
\examples{
nw <- random_network(10)
nw <- gen_partial_correlations(nw)
module <- nw$modules[[1]]
named_module <- set_module_name(module, "new name")
}
