% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{set_module_struct}
\alias{set_module_struct}
\title{Set the connections for a module}
\usage{
set_module_struct(module, struct)
}
\arguments{
\item{module}{The 'network_module' object to modify.}

\item{struct}{A vector or matrix of weights for each connetions. If a vector,
its length must equal #nodes(#nodes-1)/2, where p is the number of nodes in the module. 
If a matrix, it should be square with the number of columns equal to p; only
the upper-triangle of the matrix will be used.}
}
\value{
The modified 'network_module' object.
}
\description{
Set the connections for a module
}
