% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{set_module_weights}
\alias{set_module_weights}
\title{Set the connection weights for a module}
\usage{
set_module_weights(module, weights)
}
\arguments{
\item{module}{The 'network_module' object to modify.}

\item{weights}{A vector or matrix of weights for each connetions. If a vector,
its length must equal the number of connections in the module. If a matrix,
it should be square with the number of columns equal to the number of nodes 
in the module; only the entries in the lower triangle that correspond to 
connections in the module will be used.}
}
\value{
The modified 'network_module' object.
}
\description{
Set the connection weights for a module
}
