% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{summarize_confusion}
\alias{summarize_confusion}
\title{Provides summary measures for confusion matrix}
\usage{
summarize_confusion(confusion)
}
\arguments{
\item{confusion}{The confusion matrix.}
}
\value{
A list containing the sensitivity (i.e. true positive rate; recall), 
specificity (i.e. true negative rate), true discovery rate (i.e. positive 
predictive value; precision), true non-discovery rate (i.e. negative 
predictive value), and f1-score (geometric mean of tdr and sensitivity).
}
\description{
Provides summary measures for confusion matrix
}
