% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{trim_modules}
\alias{trim_modules}
\title{Subsets all network modules to their largest components}
\usage{
trim_modules(network)
}
\arguments{
\item{network}{The network to modify.}
}
\value{
The modified network.
}
\description{
Any nodes that are disconnected from the largest component in each module
are removed from that module. No nodes are removed from the network itself.
}
