% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{update_module_with_random_struct}
\alias{update_module_with_random_struct}
\title{Generate small-world network structure for module}
\usage{
update_module_with_random_struct(module, lattice_neig = 1,
  rewire_prob = 0.8, ...)
}
\arguments{
\item{module}{The network_module object to modify.}

\item{lattice_neig}{A positive integer used for generating small-world 
networks for modules. If the number of nodes, p, in the module is 
p >= (1 + 2 * 'lattice_neig'), then
the module will contain (p * 'lattice_neig') edges. This can be used to 
the desired level of sparsity. For sparsity in [0, 1], with one indicating no 
edges (completely sparse), setting 
'lattice_neig' = floor((1 - sparsity) * (p - 1) / 2) will constrain
the sparsity to above the desired level. Note, if sparsity is nearly 1, this
value for 'lattice_neig' may be zero, in which case p will need to be 
increased or sparsity will need to be lowered.}

\item{rewire_prob}{Used for generating small-world networks for modules; 
specifies the rewiring probability of each initial edges. At 0 the graph is
a lattice ring, and at 1 it is a random graph.}

\item{...}{Additional parameters are ignored.}
}
\value{
An updated 'network_module' object.
}
\description{
The small-world network is generated using the Watts-Strogatz method.
See ?igraph::watts.strogatz.game for details.
}
