% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{vector_to_matrix}
\alias{vector_to_matrix}
\title{Turn a vector into a matrix}
\usage{
vector_to_matrix(x, diag_val = 0)
}
\arguments{
\item{x}{the vector containing lower.tri entries of a symmetric matrix.}

\item{diag_val}{value to put along the diagonal.}
}
\value{
A symmetric matrix.
}
\description{
The input vector is assumed to be the lower.tri of a symmetric matrix. This
function recreates the original matrix from the vector.
}
