% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{add_random_module_to_network}
\alias{add_random_module_to_network}
\title{Adds a random module of a given size to the network}
\usage{
add_random_module_to_network(network, module_size, ...)
}
\arguments{
\item{network}{The 'network' object to modify.}

\item{module_size}{The size of the module to generate.}

\item{...}{Additional arguments passed into random_module().}
}
\value{
The modified 'network' object.
}
\description{
Adds a random module of a given size to the network
}
