% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{as_single_module}
\alias{as_single_module}
\title{Collapses all modules in network into a single module}
\usage{
as_single_module(network)
}
\arguments{
\item{network}{The 'network' object to modify}
}
\value{
The modified 'network' object.
}
\description{
This modification can be used if it is desired to simulate from a single
GGM rather than averaging over the GGMs for each module.
}
