% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{create_cytoscape_file}
\alias{create_cytoscape_file}
\title{Create a .csv file for Cytoscape}
\usage{
create_cytoscape_file(graph, output_name, dir = "output/")
}
\arguments{
\item{graph}{the p by p adjacency matrix (0's and 1's) of the network.}

\item{output_name}{a name for the .csv file.}

\item{dir}{(optional) output directory.}

\item{scores}{(optional) the p by p matrix of scores (weights) of each association.}

\item{gene_names}{(optional) a vector of names for the p genes in the network.}
}
\description{
In Cytopscape, use File -> Import -> Network -> File. Select the .csv file
generated by this function. There will be a popup window. The source, 
interaction, and target columns should automatically be identified. Click OK.
}
