% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoscape.R
\name{create_cytoscape_file}
\alias{create_cytoscape_file}
\title{Create a .csv file for Cytoscape}
\usage{
create_cytoscape_file(graph, output_name, dir = "output/")
}
\arguments{
\item{graph}{A 'network_plot' object. See ?plot_network().}

\item{output_name}{A name for the .csv file.}

\item{dir}{(optional) Output directory.}
}
\description{
In Cytopscape, use File -> Import -> Network -> File. Select the .csv file
generated by this function. There will be a popup window. The source, 
interaction, and target columns should automatically be identified. Click OK.
}
