% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{create_module_from_association_matrix}
\alias{create_module_from_association_matrix}
\title{Create a module from an association matrix}
\usage{
create_module_from_association_matrix(association_matrix, nodes = NULL,
  module_name = NULL)
}
\arguments{
\item{nodes}{A numeric vector indicating which nodes in the network are
contained in this module.}

\item{module_name}{(optional) Character string specifying the name of the 
module. If NULL, the module will be unnamed.}

\item{association}{matrix}
}
\value{
A 'network_module' object.
}
\description{
The edge weights in the module will be set to the corresponding values
in the association matrix. Note, only the off-diagonal elements of the 
association matrix are used.
}
