% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{create_network}
\alias{create_network}
\title{Create a network object.}
\usage{
create_network(p, n_cliques = NULL, clique_size = NULL, cliques = NULL,
  n_hubs = NULL, hub_size = NULL, hubs = NULL, n_modules = NULL,
  module_size = NULL, modules = NULL, nonoverlapping = FALSE,
  module_prob = 0.05)
}
\arguments{
\item{p}{The number of nodes in the graph}

\item{n_cliques}{The number of cliques in the graph.}

\item{clique_size}{A vector containing the size of each clique.}

\item{cliques}{(optional) A list of vectors, with each vector specifying a
set genes within a clique.}

\item{n_hubs}{The number of hub genes in the graph.}

\item{hub_size}{A vector containing the size (hub degree + 1) of each hub.}

\item{hubs}{(optional) A list of vectors, with each vector specifying a set
of hub genes; the first entry in the vector is the hub.}

\item{nonoverlapping}{If true, cliques and hubs will not share any nodes.}
}
\value{
A network object.
}
\description{
Creates a graph with certain features. This network is then
used with other sample generating methods to obtain count data.
}
\examples{
create_network(p = 50, cliques = c(1:5, 6:10), hubs = c(11:15, 16:25))
}
