% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{create_network_from_adjacency_matrix}
\alias{create_network_from_adjacency_matrix}
\title{Create a network object from adjacency matrix}
\usage{
create_network_from_adjacency_matrix(adjacency_matrix, ...)
}
\arguments{
\item{adjacency_matrix}{The adjacency matrix for the network. This is converted
to a single module structure.}

\item{...}{Additional arguments passed to
create_module_from_adjacency_matrix().}
}
\value{
A network object.
}
\description{
Create a network object from adjacency matrix
}
