% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{create_network_from_association_matrix}
\alias{create_network_from_association_matrix}
\title{Create a network object from an association matrix}
\usage{
create_network_from_association_matrix(association_matrix, ...)
}
\arguments{
\item{association_matrix}{The association matrix for the network. This is converted
to a single module structure with partial correlations specified by the
nonzero values in the matrix.}

\item{...}{Additional arguments passed to 
create_module_from_association_matrix().}
}
\value{
A network object.
}
\description{
Create a network object from an association matrix
}
