% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_zinb.R
\name{dzinb}
\alias{dzinb}
\title{The Zero-Inflated Negative Binomial Distribution}
\usage{
dzinb(x, size, mu, rho = 0, log = FALSE)
}
\arguments{
\item{x}{A vector of quantities.}

\item{size}{The dispersion paramater used in dnbinom.}

\item{mu}{The distribution mean.}

\item{rho}{The zero-inflation parameter.}

\item{log}{Logical; if TRUE, then log(d) is returned.}

\item{return}{The value(s) of the density function evaluated at x.}
}
\description{
The Zero-Inflated Negative Binomial Distribution
}
