% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_zinb.R
\name{est_params_from_reference}
\alias{est_params_from_reference}
\title{Estimate ZINB parameters from reference data}
\usage{
est_params_from_reference(reference, verbose = TRUE)
}
\arguments{
\item{reference}{Either a vector or data.frame of counts from a reference
gene expression profile. If a data.frame is provided, each column should
correspond to a gene.}

\item{verbose}{Boolean indicator for message output.}
}
\value{
Returns a list containing a matrix of parameter estimates 'size', 
'mu', and 'rho' for each gene in the reference, and the reference dataset
used.
}
\description{
The observations in the reference dataset should be as homogeneous as possible.
For example, we should not expect differential expression or differential 
connectivity of genes within the sample. If the data are heterogeneous, the
estimation of the parameters may be unreliable.
}
