% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdr.R
\name{fdr}
\alias{fdr}
\title{Determine which associations are significant}
\usage{
fdr(scores)
}
\arguments{
\item{scores}{A symmetric p by p matrix of association scores.}
}
\value{
A p by p matrix of likelihood.
}
\description{
Determines which associations are significant using emperical 
bayes and false discovery rate adjustment.
}
