% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_zinb.R
\name{gen_counts}
\alias{gen_counts}
\title{Generate RNA-seq counts}
\usage{
gen_counts(n, network, reference = NULL, verbose = TRUE)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{network}{A 'network' object or list of 'network' objects.}

\item{reference}{Either a vector or data.frame of counts from a reference
gene expression profile. If a data.frame is provided, each column should
correspond to a gene. If both 'reference' and 'params' are NULL, then parameters
are estimated from the kidney dataset.}

\item{verbose}{Boolean indicator for message output.}
}
\value{
A list containing the generated counts and the ZINB parameters used
to create them. If a list of networks were provided, then the results for
each network are returned as a list.
}
\description{
The count data are generated based on the gene-gene associations of an
udnerlying network. An association structure is imposed by first generating 
data from a multivariate Gaussian distribution, and counts are then obtained
through the inverse tranformation method. To generate realistic counts, either 
a reference dataset or parameters for the ZINB model (size, mu, rho) can be provided.
}
