% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_zinb.R
\name{gen_counts}
\alias{gen_counts}
\title{Generate samples from a zero-finlated negative-binomial distribution}
\usage{
gen_counts(n, ..., reference = NULL, params = NULL, verbose = TRUE)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{reference}{Either a vector or data.frame of counts from a reference
gene expression profile. If a data.frame is provided, each column should
correspond to a gene. If NULL, the kidney dataset is used.}

\item{params}{A matrix of ZINB parameter values; each column should contain 
the size, mu, and rho parameters for a gene.}

\item{verbose}{Boolean indicator for message output.}

\item{network}{The underlying network (from create_network())}
}
\value{
A list containing the n by p matrix of samples.
}
\description{
Generates correlated count data based on an underlying association matrix. 
The negative-binomial distribution is simulated using the gamma-Poisson mixture
model. A correlation structure is imposed by first generating data from
a multivariate Gaussian distribution, then applying the inverse CDF method. 
For reference genes with excessive zeroes, a zero-inflated model is used.
}
