% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gamma_poisson.R
\name{gen_gamma_poisson}
\alias{gen_gamma_poisson}
\title{Generate sample from overdispered Poisson}
\usage{
gen_gamma_poisson(n, network, mu, overdispersion = 1, intensity = 1,
  k = 1.5)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{network}{The underlying network (from create_network())}

\item{mu}{Mean expression for each gene. If length of mu is not equal to the number of
genes in the network, a random sample with replacement from mu will be used.}

\item{overdispersion}{A value > 0. Adjusts the amount of overdispersion in the generated counts.}

\item{intensity}{A value > 0. Used as the standard deviation of sampled edge weights.}

\item{k}{A value in [1, 2]. Adjusts the tail behavior of the distribution of generated counts.}
}
\description{
Generates count data based on the overdispered Poisson distribution.
}
