% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gamma_poisson.R
\name{gen_gamma_poisson}
\alias{gen_gamma_poisson}
\title{Generate sample from overdispered Poisson}
\usage{
gen_gamma_poisson(n, network, mu = NULL, overdispersion = 50,
  intensity = 1.5, k = 1.5)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{network}{The underlying network (from create_network())}

\item{mu}{Mean expression for each gene. If length of mu is not equal to the number of
genes in the network, a random sample with replacement from mu will be used. If NULL,
reference means are obtained from `get_reference_count_means()`.}

\item{overdispersion}{A value > 0. Adjusts the amount of overdispersion in the generated counts.}

\item{intensity}{A value > 0. Used as the standard deviation of sampled edge weights.}

\item{k}{A value in [1, 2]. Adjusts the tail behavior of the distribution of generated counts.}
}
\value{
A list containing the n by p matrix of samples, the underlying network,
and an n by p matrix containing mean expression values for each gene on each sample
after incorporating the underlying network.
}
\description{
Generates count data based on the overdispered Poisson distribution.
}
\examples{
n <- 10
network <- create_network(p = 100, modules = list(1:100)) # Create small-world network.
mu <- get_reference_count_means()
# Generate RNA-seq data using the network and reference means:
x <- gen_gamma_poisson(n, network, mu)$x
# Obtain co-expression network using cPLS algorithm with fdr rate of 0.05.
run_cpls(x, fdr = 0.05)
}
