% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_gaussian.R
\name{gen_gaussian}
\alias{gen_gaussian}
\title{Generate sample from a Gaussian graphical model.}
\usage{
gen_gaussian(n, network = NULL, sigma = NULL)
}
\arguments{
\item{n}{The number of samples to generate.}

\item{network}{The underlying network (from create_network()).}

\item{sigma}{The covariance matrix to use. If network$sigma or network$precision 
is defined, they will be used instead of sigma. Otherwise, if sigma NULL, 
a matrix is generated for sigma.}
}
\value{
A list containing the n by p matrix of samples, the underlying network,
and the p by p covariance matrix.
}
\description{
Generates count data based on the multivariate normal distribution.
}
\examples{
n <- 10
# Create a simple small-world network.
network <- create_network(p = 100, modules = list(1:100)) 
# Generate "normalized" expression data from the network:
x <- gen_gaussian(n, network)$x
}
