% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{get_adj_matrix_from_network}
\alias{get_adj_matrix_from_network}
\title{Get adjacency matrix of a network}
\usage{
get_adj_matrix_from_network(network)
}
\arguments{
\item{network}{The network to get adjacency matrix for.}
}
\value{
A p by p adjacency matrix with entry ij = 1 if node i and j are 
connected, and 0 otherwise.
}
\description{
The adjacency matrix is constructed from all the structures in the network
and summarizes which nodes are connected.
}
