% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_adjacency_matrix}
\alias{get_adjacency_matrix}
\title{Get adjacency matrix}
\usage{
get_adjacency_matrix(...)
}
\arguments{
\item{...}{Either a 'network', 'network_module', or 'matrix' object.}
}
\value{
An adjacency matrix with entry ij = 1 if node i and j are 
connected, and 0 otherwise. The diagonal entries are all zero.
}
\description{
The adjacency matrix is constructed from all modules in a network.
}
