% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_association_matrix.matrix}
\alias{get_association_matrix.matrix}
\title{Get association matrix}
\usage{
\method{get_association_matrix}{matrix}(x, tol = 10^-13, ...)
}
\arguments{
\item{x}{A 'matrix' object.}

\item{tol}{A numeric value. Associations within tol from zero are set to zero.}

\item{...}{Additional arguments.}
}
\value{
Returns the matrix with diagonal elements set to zero.
}
\description{
All off-diagonal values in the matrix are left unchanged. The diagonal
is set to zero.
}
