% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_association_matrix.network}
\alias{get_association_matrix.network}
\title{Get the association matrix of a network}
\usage{
\method{get_association_matrix}{network}(x, tol = 10^-13, ...)
}
\arguments{
\item{x}{A weighted 'network' object.}

\item{tol}{A numeric value. Associations within tol to zero are set to zero.}

\item{...}{Additional arguments.}
}
\value{
An association matrix with entry ij != 0 if node i and j have a nonzero
conditional linear association (in the Gaussian graphical model), and 0 
otherwise. The diagonal entries are all zero.
}
\description{
The association matrix for a weighted network is recovered from its
covariance structure (see ?get_sigma.network). The off-diagonal entries are 
interpreted as partial correlations, and the diagonal entries are set to zero.
}
\note{
Note that the connections in an adjacency matrix and association matrix
will likely differ. The adjacency matrix only considers direct connections in 
the network, whereas the association matrix takes into account the fact that 
there are overlapping modules which can create conditional dependencies
between two genes in seperate modules (i.e. genes that don't have a direct
connection in the graph).
}
