% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_association_matrix.network_module}
\alias{get_association_matrix.network_module}
\title{Get association matrix of a module}
\usage{
\method{get_association_matrix}{network_module}(module, ...)
}
\arguments{
\item{module}{A weighted 'network_module' object.}
}
\value{
An association matrix with entry ij != 0 if node i and j are 
connected, and 0 otherwise. The diagonal of the association matrix is set to 0.
}
\description{
Get association matrix of a module
}
