% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{get_edge_weights_from_module}
\alias{get_edge_weights_from_module}
\title{Get edge weights.}
\usage{
get_edge_weights_from_module(module)
}
\arguments{
\item{module}{The 'network_module' object to get edge weights for.}
}
\value{
A vector containing the weights of each edge. If the edges are 
unweighted, then a vector of 1's is returned. If there are no edges, in the
module, then NULL is returned.
}
\description{
Get edge weights.
}
\examples{
nw <- random_network(10)
nw <- gen_partial_correlations(nw)
module <- nw$modules[[1]]
get_edge_weights_from_module(module)
}
