% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_node_names.default}
\alias{get_node_names.default}
\title{Get node names}
\usage{
\method{get_node_names}{default}(x, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{...}{Additional arguments.}
}
\value{
A vector containing the node names or node indicies.
}
\description{
Get node names
}
