% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_node_names.network}
\alias{get_node_names.network}
\title{Get node names of a network}
\usage{
\method{get_node_names}{network}(x, ...)
}
\arguments{
\item{x}{The 'network' object to get node names from.}

\item{...}{Additional arguments.}
}
\value{
A vector containing the node names
}
\description{
Get node names of a network
}
