% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_node_names}
\alias{get_node_names}
\title{Get node names}
\usage{
get_node_names(...)
}
\arguments{
\item{...}{Either a 'network', 'network_module', or 'matrix' object.}
}
\value{
A vector containing the node names or node indicies.
}
\description{
Get node names
}
