% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{get_summary_for_node}
\alias{get_summary_for_node}
\title{Get summary for a node in the network.}
\usage{
get_summary_for_node(node, network)
}
\arguments{
\item{node}{The node to summarize. Can be a character string 
corresponding to a name of a node in the network, or an integer value from 
1 to p corresponding to the index of a node.}

\item{network}{A network object.}
}
\value{
A list containing summary, a list of strings detailing each 
structure the node belongs to; struct_count, the number of structures
the node belongs to; and degree, the total number of connections to the 
node (summed over all structures).
}
\description{
Get summary for a node in the network.
}
