% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{is_weighted.network}
\alias{is_weighted.network}
\title{Check if a network is weighted}
\usage{
\method{is_weighted}{network}(network, ...)
}
\arguments{
\item{network}{The 'network' object to check.}
}
\value{
A boolean value that is TRUE if all of the modules in the network
are weighted and FALSE otherwise. If the network contains no modules or
no connections, then this function returns TRUE.
}
\description{
Check if a network is weighted
}
