% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot.network}
\alias{plot.network}
\title{Plot function for network class}
\usage{
\method{plot}{network}(network, compare_graph = NULL, as_subgraph = FALSE,
  node_scale = 5, edge_scale = 1, coords = NULL, main = "Untitled",
  include_vertex_labels = TRUE, ...)
}
\arguments{
\item{network}{Either a network object or adjacency matrix of the network.}

\item{compare_graph}{The plot of another network to use for comparison.}

\item{as_subgraph}{If true, only nodes of positive degree will be shown.}

\item{node_scale}{Used for scaling of nodes.}

\item{edge_scale}{Used for scaling of edges.}

\item{coords}{Layout used for the network.}

\item{main}{A string containing the title for the graph.}

\item{include_vertex_labels}{If true, the verticies will be labeled.}
}
\value{
Creates a plot of the network and returns a graph object. 
The graph object can be passed back into a future call of plot.network() 
through the `compare_edge` argument, which will setup the plot for easier 
comparison between the old graph and the graph of `network`.
}
\description{
This function plots the given network. If the result of another plot is 
provided, this plot will be modified for easier comparison.
}
