% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot.network}
\alias{plot.network}
\title{Plot function for network class}
\usage{
\method{plot}{network}(network, compare_graph = NULL, weighted = NULL,
  as_subgraph = FALSE, node_scale = 5, edge_scale = 1,
  coords = layout.fruchterman.reingold, main = "Untitled", ...)
}
\arguments{
\item{network}{Either a network object or adjacency matrix of the network.}

\item{compare_graph}{The plot of another network to use for comparison.}

\item{weighted}{Are the edges weighted?}

\item{as_subgraph}{If true, only nodes of positive degree will be shown.}

\item{node_scale}{Used for scaling of nodes.}

\item{edge_scale}{used for scaling of edges.}

\item{coords}{Layout used for the network.}
}
\description{
This function plots the given network. If the result of another plot is 
provided, this plot will be modified for easier comparison.
}
