% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.network_module}
\alias{plot.network_module}
\title{Plot function for 'network_module' object.}
\usage{
\method{plot}{network_module}(x, ...)
}
\arguments{
\item{x}{A 'network_module' object.}

\item{...}{Additional arguments passed to plot_network().}
}
\value{
Creates a plot of the module and returns a graph object. 
See ?plot_network for details.
}
\description{
Plot function for 'network_module' object.
}
\examples{
module <- random_module(1:10)
plot(module)
}
