% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.network_plot}
\alias{plot.network_plot}
\title{Plot function for 'network_plot' class}
\usage{
\method{plot}{network_plot}(network_plot,
  vertex.color = network_plot$vertex.color, vertex.label.font = 2,
  vertex.size = network_plot$vertex.size,
  vertex.frame.color = network_plot$vertex.frame.color,
  vertex.label.color = network_plot$vertex.label.color,
  vertex.label.cex = 0.7, edge.color = network_plot$edge.color,
  edge.width = network_plot$edge.width, layout = network_plot$coords,
  ...)
}
\arguments{
\item{network_plot}{A 'network_plot' object obtained from plot.network() or
plot_network().}

\item{...}{Additional arguments passed to plot.igraph().}
}
\description{
Plot function for 'network_plot' class
}
