% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_modules}
\alias{plot_modules}
\title{Visualize a network and its modules}
\usage{
plot_modules(network, compare_graph = NULL, as_subgraph = TRUE,
  node_scale = 5, edge_scale = 2, node_color = adjustcolor("orange",
  0.5), group_color = RColorBrewer::brewer.pal(min(9, max(3,
  length(network$modules))), "Set1"), coords = NULL,
  generate_layout = igraph::nicely, include_vertex_labels = TRUE,
  show_legend = ifelse(length(network$modules) <= 10, TRUE, FALSE),
  legend_position = "topright", legend_horizontal = FALSE,
  display_plot = TRUE, ...)
}
\arguments{
\item{network}{A 'network' object to plot.}

\item{compare_graph}{The plot of another network to use for comparison.}

\item{as_subgraph}{If TRUE, only nodes of positive degree will be shown.}

\item{node_scale}{Used for scaling of nodes.}

\item{edge_scale}{Used for scaling of edges.}

\item{node_color}{The color used for the nodes.}

\item{group_color}{A vector of colors used for the modules.}

\item{include_vertex_labels}{If TRUE, the verticies will be labeled.}

\item{show_legend}{If TRUE, a legend for the modules - module colors is shown.
Defaults to TRUE if there are 10 or fewer modules and to FALSE otherwise.}

\item{legend_position}{The location of the legend. Can be any one of "bottomright",
"bottom", "bottomleft", "left", "topleft", "top", "topright", "right" or "center".}

\item{legend_horizontal}{If TRUE, the legend will be displayed horizontally.}

\item{display_plot}{If TRUE (default), the plot will be generated and displayed.}

\item{...}{Additional arguments passed to plot.igraph().}

\item{generate_coords}{A function to generate the layout of a graph; used
if coords is NULL. See ?igraph::layout_ for details. Other options include 
'igraph::as_star', 'igraph::in_circle', and 'igraph::with_fr', among many others.}
}
\value{
A 'network_plot' object for the network. This object can be passed 
back into a future call of plot.network() through the `compare_graph` 
argument, which will setup the plot for easier comparison between the old 
graph and the new graph of `network`.
}
\description{
This function plots a network and highlights the individual modules. 
An attempt is made to layout the nodes so that any visual overlaps among modules
correspond to true overlaps in the network, however it is possible that 
a node may appear to be in multiple modules in the visualization when it does
not actually belong to multiple modules. If the result of another plot is 
provided using the 'compare_graph' argument, then the layout of this network
will be based on that plot and convex hulls are drawn to trace out the modules; 
in this case it is likely that the displayed modules will contain extraneous
nodes.
}
