% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network_diff}
\alias{plot_network_diff}
\title{Plot function for network class}
\usage{
plot_network_diff(network_1, network_2, weighted = NULL, as_subgraph = TRUE,
  node_scale = 5, edge_scale = 1,
  coords = igraph::layout.fruchterman.reingold, main = "Untitled", ...)
}
\arguments{
\item{weighted}{Are the edges weighted?}

\item{as_subgraph}{If true, only nodes of positive degree will be shown.}

\item{node_scale}{Used for scaling of nodes.}

\item{edge_scale}{used for scaling of edges.}

\item{coords}{Layout used for the network.}

\item{network}{Either a network object or adjacency matrix of the network.}

\item{compare_graph}{The plot of another network to use for comparison.}
}
\value{
Creates a plot of the network and returns a graph object. 
The graph object can be passed back into a future call of plot.network() 
through the `compare_edge` argument, which will setup the plot for easier 
comparison between the old graph and the graph of `network`.
}
\description{
This function plots the given network. If the result of another plot is 
provided, this plot will be modified for easier comparison.
}
