% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_network_diff}
\alias{plot_network_diff}
\title{Plotting to compare two networks}
\usage{
plot_network_diff(network_1, network_2, as_subgraph = FALSE,
  compare_graph = NULL, node_scale = 5, edge_scale = 2,
  node_color = adjustcolor("orange", 0.5),
  generate_layout = igraph::nicely, include_vertex_labels = TRUE, ...)
}
\arguments{
\item{network_1}{Either a network object or adjacency matrix of the network.}

\item{network_2}{Either a network object or adjacency matrix of the network.}

\item{as_subgraph}{If TRUE, only nodes of positive degree will be shown.}

\item{compare_graph}{The plot of another network to use for comparison.}

\item{node_scale}{Used for scaling of nodes.}

\item{edge_scale}{Used for scaling of edges.}

\item{node_color}{The color used for the nodes.}

\item{include_vertex_labels}{If TRUE, the verticies will be labeled.}

\item{...}{Additional arguments passed to plot.igraph().}

\item{generate_coords}{A function to generate the layout of a graph; used
if coords is NULL. See ?igraph::layout_ for details. Other options include 
'igraph::as_star', 'igraph::in_circle', and 'igraph::with_fr', among many others.}

\item{display_plot}{If TRUE (default), the plot will be generated and displayed.}
}
\value{
Creates a plot of the network and returns a graph object. 
The graph object can be passed back into a future call of plot.network() 
through the `compare_edge` argument, which will setup the plot for easier 
comparison between the old graph and the graph of `network`.
}
\description{
This function plots the overlap of two networks. Black edges are shared by
both networks, tan edges are in network_1 but not network_2, and red edges
are in network_2 but not network_1.
}
