% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_network_sim}
\alias{plot_network_sim}
\title{Plot the similarity between two networks}
\usage{
plot_network_sim(network_1, network_2, compare_graph, ...)
}
\arguments{
\item{network_1}{A weighted 'network' or 'matrix' object.}

\item{network_2}{A weighted 'network' or 'matrix' object.}

\item{compare_graph}{The plot of another network to use for comparison.}

\item{...}{Additional arguments passed to 'plot_network()'.}
}
\value{
Creates a plot of the network and returns a graph object. 
The graph object can be passed back into a future call of 'plot_network()',
'plot_network_diff()' or 'plot_network_sim()'
through the 'compare_edge' argument, which will setup the plot for easier 
comparison between the old graph and the graph of 'network'.
}
\description{
This function plots the similarity of connections between two networks. 
Both networks must be weighted. The width of each edge corresponds to 
the strength of similarity and is calculated by exp(0.5(log(|s1 + s2|) +
log(|s1|) + log(|s2|))), where s1 and s2 is the strength of a particular
connection in network_1 and network_2, respectively
}
