% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{print.network_plot}
\alias{print.network_plot}
\title{Print function for 'network_plot' class}
\usage{
\method{print}{network_plot}(x, ...)
}
\arguments{
\item{x}{A 'network_plot' object obtained from plot.network() or
plot_network().}

\item{...}{Additional arguments passed to plot().}
}
\description{
Displays the network plot.
}
