% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_zinb.R
\name{pzinb}
\alias{pzinb}
\title{The Zero-Inflated Negative Binomial Distribution}
\usage{
pzinb(q, size, mu, rho, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{A vector of quantities.}

\item{size}{The dispersion paramater used in dnbinom.}

\item{mu}{The distribution mean.}

\item{rho}{The zero-inflation parameter.}

\item{lower.tail}{Logical; if TRUE, then probabilities are P(X <= x).
Otherwise, P(X > x).}

\item{log.p}{Logical; if TRUE, then log(p) is returned.}
}
\description{
The Zero-Inflated Negative Binomial Distribution
}
