% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{random_module}
\alias{random_module}
\title{Create a random module}
\usage{
random_module(nodes, module_name = NULL, add_weights = FALSE, ...)
}
\arguments{
\item{nodes}{A numeric vector indicating which nodes in the network 
are contained in this module.}

\item{module_name}{(optional) Character string specifying the name of the 
module. If NULL, the module will be unnamed.}

\item{add_weights}{If TRUE, weights will be generated for the module 
connections.}

\item{...}{Additional arguments passed to 'update_module_with_random_edges()' 
and 'update_module_with_random_weights()'.}
}
\value{
A 'network_module' object.
}
\description{
Create a random module
}
\details{
See ?igraph::watts.strogatz.game for details on generating
small-world networks.
}
