% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{random_network}
\alias{random_network}
\title{Create a network object.}
\usage{
random_network(p, n_modules = ceiling(p * 0.05), min_module_size = 10,
  max_module_size = Inf, avg_module_size = 25, sd_module_size = 20,
  repeated_module_factor = 0.2, ...)
}
\arguments{
\item{p}{The number of nodes in the network; 'p' is required to be 
between 10 and 20000.}

\item{n_modules}{The number of modules to include in the network. The
default value is 5% of the number of nodes in the network.}

\item{min_module_size}{The minimum number of nodes in a module.}

\item{max_module_size}{(optional) If specified, any generated module sizes 
above this value will be reduced to 'max_module_size'.}

\item{avg_module_size}{The average number of nodes in a module.}

\item{sd_module_size}{The standard deviation of module size.}

\item{repeated_module_factor}{A value between 0 and 1. A gene selected for a 
module has its probability of being selected for another module multiplied by 
this factor.}
}
\value{
An unweighted network object.
}
\description{
Creates an unweighted 'network' object containing randomly generated 
modules.
}
