% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reference}
\alias{reference}
\title{RNA-seq reference dataset}
\format{A list containing two data frames:
\describe{
  \item{$rnaseq}{A 1093 by 15944 data frame containing the raw RNA-seq 
  expression counts}
  \item{$params}{A 3 by 15944 data frame containing the estimated 
  ZINB parameters for each expression profile}
}}
\source{
\url{http://www.linkedomics.org/data_download/TCGA-BRCA/}
}
\usage{
reference
}
\description{
The reference is a breast invasive carcinoma dataset containing gene 
expression profiles generated by The Cancer Genome Atlas (TCGA) and 
downloaded using the LinkedOmics portal. The dataset contains 1093 samples 
and 15944 genes. The reference is a list containing a data frame of the 
expression data and a data frame of estimated ZINB parameters for each 
expression profile.
}
\keyword{datasets}
