% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{remove_connections_to_node.default}
\alias{remove_connections_to_node.default}
\title{Remove connections to a node.}
\usage{
\method{remove_connections_to_node}{default}(...)
}
\arguments{
\item{...}{A 'network', 'network_module', or 'matrix' object to modify.}
}
\value{
The modified object.
}
\description{
Remove connections to a node.
}
