% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{remove_connections_to_node.network_module}
\alias{remove_connections_to_node.network_module}
\title{Remove connections to a node.}
\usage{
\method{remove_connections_to_node}{network_module}(module, node,
  prob_remove, weights = NULL, exponent = 0)
}
\arguments{
\item{module}{A 'network_module' object to modify.}

\item{node}{The node to unwire.}

\item{prob_remove}{A value between 0 and 1. Each connection to 'node_index' 
will be removed with probability equal to 'remove_prob'.}

\item{weights}{(Optional) A vector of weights for each node. These are used
in addition to the degree of each node when sampling neighbors to unwire from.}

\item{exponent}{The exponent used for weighted sampling. When exponent = 0,
neighboring nodes are sampled uniformly. When exponent > 0, the sampling 
probability is based on node weights.}
}
\value{
The modified module.
}
\description{
Remove connections to a node.
}
