% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{remove_connections_to_node}
\alias{remove_connections_to_node}
\title{Remove connections to a node.}
\usage{
remove_connections_to_node(x, ...)
}
\arguments{
\item{x}{A 'network', 'network_module', or 'matrix' object to modify.}

\item{...}{Additional parameters.}
}
\value{
The modified object.
}
\description{
Remove connections to a node.
}
