% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{remove_connections_to_node}
\alias{remove_connections_to_node}
\title{Remove all connections to a node}
\usage{
remove_connections_to_node(node, network)
}
\arguments{
\item{node}{The node to cut connections from. Can be a character string 
corresponding to a name of a node in the network, or an integer value from 
1 to p corresponding to the index of a node.}

\item{network}{The network to modify.}
}
\value{
The modified network.
}
\description{
Remove all connections to a node
}
