% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{remove_small_components_from_module}
\alias{remove_small_components_from_module}
\title{Subset module onto largest component}
\usage{
remove_small_components_from_module(module)
}
\arguments{
\item{module}{The 'network_module' object to modify.}
}
\value{
The modified 'network_module' object.
}
\description{
Any nodes that are disconnected from the largest component in the module
are removed.
}
