% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{remove_weights.matrix}
\alias{remove_weights.matrix}
\title{Removes the weights of all connections}
\usage{
\method{remove_weights}{matrix}(m, ...)
}
\arguments{
\item{m}{A 'matrix' object.}
}
\value{
The modified object.
}
\description{
Removes the weights of all connections
}
