% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{remove_weights}
\alias{remove_weights}
\title{Removes the weights of all connections}
\usage{
remove_weights(x, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{...}{Additional arguments.}
}
\value{
The modified object.
}
\description{
Removes the weights of all connections
}
