% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rewire_connections.matrix}
\alias{rewire_connections.matrix}
\title{Rewire connections.}
\usage{
\method{rewire_connections}{matrix}(x, prob_rewire, weights = NULL,
  exponent = 0, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{prob_rewire}{A value between 0 and 1. The connections to each node 
will be rewired with probability equal to 'prob_rewire'.}

\item{weights}{(Optional) A vector of weights for each node. These are used
in addition to the degree of each node when sampling a node to rewire to.}

\item{exponent}{The exponent used for weighted sampling. When exponent = 0,
nodes are sampled uniformly. When exponent > 0, the sampling probability
is based on node weights.}

\item{...}{Additional arguments.}
}
\value{
The modified object x.
}
\description{
Rewire connections.
}
