% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{rewire_connections.network_module}
\alias{rewire_connections.network_module}
\title{Rewire connections.}
\usage{
\method{rewire_connections}{network_module}(x, prob_rewire,
  weights = NULL, alpha = 100, beta = 1, epsilon = 10^-5, ...)
}
\arguments{
\item{x}{A 'network_module' object.}

\item{prob_rewire}{A value between 0 and 1. The connections to each node 
will be rewired with probability equal to 'prob_rewire'.}

\item{weights}{(Optional) A vector of weights for each node. These are used
in addition to the degree of each node when sampling a node to rewire to.}

\item{...}{Additional arguments.}

\item{eta}{The exponent used for weighted sampling. When eta = 0,
nodes are sampled uniformly. When eta > 0, the sampling probability
is based on node weights.}
}
\value{
The modified module.
}
\description{
Rewire connections.
}
