% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rewire_connections_to_node.default}
\alias{rewire_connections_to_node.default}
\title{Rewire connections to a node.}
\usage{
\method{rewire_connections_to_node}{default}(x, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{...}{Additional arguments.}
}
\value{
The modified object.
}
\description{
Rewire connections to a node.
}
