% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rewire_connections_to_node.matrix}
\alias{rewire_connections_to_node.matrix}
\title{Rewire connections to a node.}
\usage{
\method{rewire_connections_to_node}{matrix}(x, node, prob_rewire,
  weights = NULL, exponent = 0, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{node}{The node to rewire.}

\item{prob_rewire}{A value between 0 and 1. Each connection to 'node' 
will be rewired with probability equal to 'prob_rewire'. Note, the degree of 
'node' is unchanged after this operation.}

\item{weights}{(Optional) A vector of weights for each node. These are used
in addition to the degree of each node when sampling nodes to rewire.}

\item{exponent}{The exponent used for weighted sampling. When exponent = 0,
nodes are sampled uniformly. When exponent > 0, the sampling probability
is based on node weights.}

\item{...}{Additional arguments.}
}
\value{
The modified object m.
}
\description{
Rewire connections to a node.
}
