% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_network.R
\name{rewire_connections_to_node}
\alias{rewire_connections_to_node}
\title{Rewire node in the network.}
\usage{
rewire_connections_to_node(network, node, p = 0.2)
}
\arguments{
\item{network}{The network to modify.}

\item{node}{The node to rewire. Can be a character string if the nodes
are labeled, or a integer value from 1 to p.}

\item{p}{The probabiliy of an edge to be created.}
}
\value{
The modified network.
}
\description{
Within each module the node is a member of, all connections are reset. 
New connections are made with a specified probablility.
}
