% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{rewire_connections_to_node}
\alias{rewire_connections_to_node}
\title{Rewire connections to a node.}
\usage{
rewire_connections_to_node(x, node, ...)
}
\arguments{
\item{x}{Either a 'network', 'network_module', or 'matrix' object.}

\item{node}{The node to rewire.}

\item{...}{Additional arguments.}
}
\value{
The modified object.
}
\description{
Rewire connections to a node.
}
