% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr.R
\name{run_corr}
\alias{run_corr}
\title{Wrapper for correlation co-expression}
\usage{
run_corr(x, threshold = 0.9)
}
\arguments{
\item{x}{The n by p matrix of counts}

\item{threshold}{Cutoff for significant associations}
}
\value{
A list containing `scores`, a p by p matrix of correlations, and 
`adj_matrix`, a p by p adjacency matrix.
}
\description{
Conducts co-expression analysis using correlation for association measure.
}
