% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpls.R
\name{run_cpls}
\alias{run_cpls}
\title{Combined partial least squares with negative-binomial}
\usage{
run_cpls(x, v = 3, threshold = NULL, parallel = TRUE)
}
\arguments{
\item{x}{The n by p matrix of raw count data. Rows should correspond to
samples, and columns should correspond to genes.}

\item{v}{The number of PLS components to compute.}

\item{threshold}{(optional) the signifiance threshold for inferring the 
network through empirical bayes fdr. If NULL, only scores are returned.}

\item{parallel}{Should cpls be run in parallel? (Unix system required.)}
}
\value{
a matrix of raw association scores.
}
\description{
Reconstructs a full genomic association network from a set of 
next-generation sequencing count data.
}
