% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpls.R
\name{run_cpls_parallel}
\alias{run_cpls_parallel}
\title{Combined partial least squares with negative-binomial}
\usage{
run_cpls_parallel(x, threshold = NULL, v = 1, parallel = FALSE)
}
\arguments{
\item{x}{The n by p matrix of raw count data. Rows should correspond to
samples, and columns should correspond to genes.}

\item{threshold}{(optional) the fdr signifiance threshold for inferring the
network through empirical bayes fdr. If NULL, all scores are returned.
Otherwise, scores with fdr above this threshold are set to zero.}

\item{v}{The number of PLS components to compute.}

\item{parallel}{Should cpls be run in parallel? (Unix system required.)}
}
\value{
a matrix of raw association scores.
}
\description{
Parallel version of run_cpls(). Reconstructs a full genomic association
network from a set of next-generation sequencing count data.
}
