% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso.R
\name{run_glasso}
\alias{run_glasso}
\title{Wrapper for glasso method}
\usage{
run_glasso(x, method = c("glasso", "mb", "ct"), criterion = c("ric",
  "stars"), verbose = FALSE, ...)
}
\arguments{
\item{x}{The n by p matrix of counts.}

\item{method}{Graph estimation method. Should be either "mb", "glasso", or "ct".
Default is "glasso".}

\item{criterion}{Model selection criterion. Should be either "ric" or "stars".
Default is "ric".}

\item{verbose}{boolean indicating whether trace information during estimation 
and selection should be printed. Default is FALSE.}
}
\value{
A list containing the p by p matrix of association scores, the 
estimation method used, and the selection criterion used.
}
\description{
Conducts co-expression analysis using glasso.
}
