% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icorr.R
\name{run_icorr}
\alias{run_icorr}
\title{Wrapper for inverse correlation matrix}
\usage{
run_icorr(x, threshold = NULL, method = "pearson")
}
\arguments{
\item{x}{The n by p matrix of counts.}

\item{threshold}{Cutoff for significant associations. If NULL, all correlations
are returned. Otherwise, correlations of magnitude at or below this threshold are 
set to zero.}

\item{method}{The method used to compute correlations. Should be either "pearson"
or "spearman". The default is Spearman, which provides 
the more conservative estimation of associations.}
}
\value{
A list containing the p by p matrix of partial correlations,
the threshold used to determine significant associations, and
the method used to compute the correlations.
}
\description{
Conducts co-expression analysis using partial correlations for association 
measure, computed by inverting the correlation matrix.
}
