% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{run_wgcna}
\alias{run_wgcna}
\title{Wrapper for WGCNA method}
\usage{
run_wgcna(x, threshold = 0, beta = 6)
}
\arguments{
\item{x}{The n by p matrix of counts.}

\item{threshold}{Cutoff for significant associations. Scores are rounded to
two decimal places. If threshold is provided, rounded scores at or below this 
threshold are set to zero. Otherwise, all scores are returned.}

\item{beta}{Tuning parameter for WGCNA.}
}
\value{
A list containing the p by p matrix of association scores, and the 
threshold used to determine significant associations.
}
\description{
Conducts co-expression analysis using WGCNA method. Preprocessing is performed
on x: the trimmed mean of M-values (TMM) normalization, which accounts 
for between-sample biases, and counts per million (CPM) scaling, which 
addresses any difference in library sizes.
}
