% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgcna.R
\name{run_wgcna}
\alias{run_wgcna}
\title{Wrapper for WGCNA method}
\usage{
run_wgcna(x, threshold = NULL, method = "pearson")
}
\arguments{
\item{x}{The n by p matrix of counts.}

\item{threshold}{Cutoff for significant associations. If NULL, all scores
are returned. Otherwise, scores at or below this threshold are set to zero.}

\item{method}{The method used to compute correlations. Should be either "pearson"
or "spearman". The default is Spearman, which provides 
the more conservative estimation of associations.}
}
\value{
A list containing the p by p matrix of association scores, and the 
threshold used to determine significant associations.
}
\description{
Conducts co-expression analysis using WGCNA method. Preprocessing is performed
on x: the trimmed mean of M-values (TMM) normalization, which accounts 
for between-sample biases, and counts per million (CPM) scaling, which 
addresses any difference in library sizes.
}
