% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_zinb.R
\name{rzinb}
\alias{rzinb}
\title{The Zero-Inflated Negative Binomial Distribution}
\usage{
rzinb(n, size, mu, rho)
}
\arguments{
\item{n}{The number of random values to return.}

\item{size}{The dispersion paramater used in dnbinom.}

\item{mu}{The distribution mean.}

\item{rho}{The zero-inflation parameter.}
}
\description{
The Zero-Inflated Negative Binomial Distribution
}
