% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networks.R
\name{sample_module_nodes}
\alias{sample_module_nodes}
\title{Sample nodes for new module}
\usage{
sample_module_nodes(n, nodes, degree, nu = 0.01, ...)
}
\arguments{
\item{n}{The number of nodes to sample.}

\item{nodes}{The nodes available to sample from.}

\item{degree}{The degree of each node.}

\item{nu}{Multiplier for nodes that are already in one or more modules.}
}
\value{
A vector of selected nodes of length m.
}
\description{
Sample nodes for new module
}
