% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_measures.R
\name{summarize_confusion}
\alias{summarize_confusion}
\title{Provides summary measures for confusion matrix}
\usage{
summarize_confusion(confusion)
}
\arguments{
\item{confusion}{The confusion matrix. This can either be an object of class
"CONFUSION" or a 2 by 2 integer matrix.}
}
\value{
A list containing several summary measures: 
sensitivity (a.k.a. true positive rate; recall), 
specificity (a.k.a. true negative rate), 
TDR (true discovery rate; a.k.a. positive predictive value; precision), 
TNDR (true non-discovery rate; a.k.a. negative predictive value), 
f1-score (harmonic mean of TDR and sensitivity), and
MCC (Matthew's correlation coefficient).
}
\description{
Provides summary measures for confusion matrix
}
